Move PythonEnvironments to the component "api" module.
